requirements.txt
amqp==5.3.1
argon2-cffi-bindings==25.1.0
billiard==4.2.1
blinker==1.9.0
brotli==1.2.0
celery==5.5.3
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
curl_cffi==0.13.0
Flask==3.1.2
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
MarkupSafe==3.0.2
mutagen==1.47.0
neat-python==0.92
packaging==25.0
prompt_toolkit==3.0.52
pycparser==2.23
pycryptodomex==3.23.0
pygame-ce==2.5.7
python-dateutil==2.9.0.post0
redis==5.2.1
requests==2.32.5
setuptools==80.9.0
six==1.17.0
tzdata==2025.2
Unidecode==1.4.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
yt-dlp==2025.12.8
yt-dlp-ejs==0.3.2
zope.event==6.0
zope.interface==8.0
